+++ node/mdSchema.ts
@@ -8,7 +8,7 @@ export const ORGANIZATION_REQUEST_FIELDS = [
   'notes',
   'created',
 ]
-export const ORGANIZATION_REQUEST_SCHEMA_VERSION = 'v0.0.2'
+export const ORGANIZATION_REQUEST_SCHEMA_VERSION = 'v0.0.3'
 
 export const ORGANIZATION_DATA_ENTITY = 'organizations'
 export const ORGANIZATION_FIELDS = [
@@ -21,7 +21,7 @@ export const ORGANIZATION_FIELDS = [
   'status',
   'created',
 ]
-export const ORGANIZATION_SCHEMA_VERSION = 'v0.0.4'
+export const ORGANIZATION_SCHEMA_VERSION = 'v0.0.5'
 
 export const COST_CENTER_DATA_ENTITY = 'cost_centers'
 export const COST_CENTER_FIELDS = ['id', 'name', 'addresses', 'organization']
@@ -56,6 +56,7 @@ export const schemas = [
         created: {
           type: 'string',
           title: 'Created',
+          format: 'date-time',
         },
       },
       'v-indexed': ['name', 'status', 'created'],
@@ -95,6 +96,7 @@ export const schemas = [
         created: {
           type: 'string',
           title: 'Created',
+          format: 'date-time',
         },
       },
       'v-indexed': ['name', 'status'],